pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-governance-platform"
version = "0.1.0"
description = "AI governance control plane for enterprise LLM deployments"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "sqlalchemy==2.0.23",
    "alembic==1.12.1",
    "psycopg2-binary==2.9.9",
    "redis==5.0.1",
    "kafka-python==2.0.2",
    "httpx==0.25.2",
    "aiohttp==3.9.1",
    "grpcio==1.59.3",
    "grpcio-tools==1.59.3",
    "protobuf==4.25.1",
    "openai==1.3.7",
    "anthropic==0.7.8",
    "tiktoken==0.5.2",
    "python-dotenv==1.0.0",
    "pyyaml==6.0.1",
    "structlog==23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "black==23.11.0",
    "ruff==0.1.6",
    "mypy==1.7.1",
]

[tool.hatch.build.targets.wheel]
packages = [
    "gateway",
    "policy_engine",
    "audit",
    "hitl",
    "model_router",
    "policies",
    "config",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

